requirements.txt
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
Flask==0.12.2
google-api-core==0.1.1
google-api-python-client==1.6.4
google-auth==1.2.1
google-auth-httplib2==0.0.3
google-cloud-pubsub==0.29.2
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.7.3
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauth2client==4.1.2
protobuf==3.5.0.post1
psutil==5.4.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
requests==2.18.4
requirements==0.1
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.13
